+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/panel/DefaultBorderLayoutPanelBuilder.java
@@ -28,17 +28,18 @@ import org.jspecify.annotations.Nullable;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
 import java.awt.BorderLayout;
+import java.util.function.Supplier;
 
 import static java.util.Objects.requireNonNull;
 
 final class DefaultBorderLayoutPanelBuilder extends AbstractComponentBuilder<Void, JPanel, BorderLayoutPanelBuilder> implements BorderLayoutPanelBuilder {
 
 	private BorderLayout layout = Layouts.borderLayout();
-	private @Nullable JComponent centerComponent;
-	private @Nullable JComponent northComponent;
-	private @Nullable JComponent southComponent;
-	private @Nullable JComponent eastComponent;
-	private @Nullable JComponent westComponent;
+	private @Nullable Supplier<? extends JComponent> centerComponent;
+	private @Nullable Supplier<? extends JComponent> northComponent;
+	private @Nullable Supplier<? extends JComponent> southComponent;
+	private @Nullable Supplier<? extends JComponent> eastComponent;
+	private @Nullable Supplier<? extends JComponent> westComponent;
 
 	DefaultBorderLayoutPanelBuilder() {}
 
@@ -50,6 +51,11 @@ final class DefaultBorderLayoutPanelBuilder extends AbstractComponentBuilder<Voi
 
 	@Override
 	public BorderLayoutPanelBuilder add(JComponent component, String constraints) {
+		return add(() -> requireNonNull(component), constraints);
+	}
+
+	@Override
+	public BorderLayoutPanelBuilder add(Supplier<? extends JComponent> component, String constraints) {
 		switch (requireNonNull(constraints)) {
 			case BorderLayout.NORTH:
 				return northComponent(component);
@@ -66,30 +72,55 @@ final class DefaultBorderLayoutPanelBuilder extends AbstractComponentBuilder<Voi
 
 	@Override
 	public BorderLayoutPanelBuilder centerComponent(JComponent centerComponent) {
+		return centerComponent(() -> requireNonNull(centerComponent));
+	}
+
+	@Override
+	public BorderLayoutPanelBuilder centerComponent(Supplier<? extends JComponent> centerComponent) {
 		this.centerComponent = requireNonNull(centerComponent);
 		return this;
 	}
 
 	@Override
 	public BorderLayoutPanelBuilder northComponent(JComponent northComponent) {
+		return northComponent(() -> requireNonNull(northComponent));
+	}
+
+	@Override
+	public BorderLayoutPanelBuilder northComponent(Supplier<? extends JComponent> northComponent) {
 		this.northComponent = requireNonNull(northComponent);
 		return this;
 	}
 
 	@Override
 	public BorderLayoutPanelBuilder southComponent(JComponent southComponent) {
+		return southComponent(() -> requireNonNull(southComponent));
+	}
+
+	@Override
+	public BorderLayoutPanelBuilder southComponent(Supplier<? extends JComponent> southComponent) {
 		this.southComponent = requireNonNull(southComponent);
 		return this;
 	}
 
 	@Override
 	public BorderLayoutPanelBuilder eastComponent(JComponent eastComponent) {
+		return eastComponent(() -> requireNonNull(eastComponent));
+	}
+
+	@Override
+	public BorderLayoutPanelBuilder eastComponent(Supplier<? extends JComponent> eastComponent) {
 		this.eastComponent = requireNonNull(eastComponent);
 		return this;
 	}
 
 	@Override
 	public BorderLayoutPanelBuilder westComponent(JComponent westComponent) {
+		return westComponent(() -> requireNonNull(westComponent));
+	}
+
+	@Override
+	public BorderLayoutPanelBuilder westComponent(Supplier<? extends JComponent> westComponent) {
 		this.westComponent = requireNonNull(westComponent);
 		return this;
 	}
@@ -98,19 +129,19 @@ final class DefaultBorderLayoutPanelBuilder extends AbstractComponentBuilder<Voi
 	protected JPanel createComponent() {
 		JPanel component = new JPanel(layout);
 		if (centerComponent != null) {
-			component.add(centerComponent, BorderLayout.CENTER);
+			component.add(centerComponent.get(), BorderLayout.CENTER);
 		}
 		if (northComponent != null) {
-			component.add(northComponent, BorderLayout.NORTH);
+			component.add(northComponent.get(), BorderLayout.NORTH);
 		}
 		if (southComponent != null) {
-			component.add(southComponent, BorderLayout.SOUTH);
+			component.add(southComponent.get(), BorderLayout.SOUTH);
 		}
 		if (eastComponent != null) {
-			component.add(eastComponent, BorderLayout.EAST);
+			component.add(eastComponent.get(), BorderLayout.EAST);
 		}
 		if (westComponent != null) {
-			component.add(westComponent, BorderLayout.WEST);
+			component.add(westComponent.get(), BorderLayout.WEST);
 		}
 
 		return component;