+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/panel/PanelBuilder.java
@@ -27,6 +27,7 @@ import javax.swing.JPanel;
 import java.awt.Component;
 import java.awt.LayoutManager;
 import java.util.Collection;
+import java.util.function.Supplier;
 
 /**
  * Builds a JPanel instance.
@@ -55,11 +56,26 @@ public interface PanelBuilder extends ComponentBuilder<Void, JPanel, PanelBuilde
 
 	/**
 	 * @param component the component to add
+	 * @return this builder instance
+	 * @see JPanel#add(Component)
+	 */
+	PanelBuilder add(Supplier<? extends JComponent> component);
+
+	/**
+	 * @param component the component to add
 	 * @param constraints the layout constraints
 	 * @return this builder instance
 	 * @see JPanel#add(Component, Object)
 	 */
 	PanelBuilder add(JComponent component, Object constraints);
+
+	/**
+	 * @param component the component to add
+	 * @param constraints the layout constraints
+	 * @return this builder instance
+	 * @see JPanel#add(Component, Object)
+	 */
+	PanelBuilder add(Supplier<? extends JComponent> component, Object constraints);
 
 	/**
 	 * @param components the components to add